requirements.txt
asgiref==3.8.1
Django>=3.2,<4.0
djangorestframework>=3.12.0
drf-yasg>=1.20.0
gunicorn>=20.0.4
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv
sqlparse==0.5.3
tzdata==2025.2
whitenoise>=5.2.0
# Add any other dependencies your project needs
